pyproject.toml
[project]
name = "crop-disease-detector-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "pillow>=12.2.0",
    "python-dotenv>=1.0.0",
    "torch>=2.11.0",
    "transformers>=5.5.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
